+++ smartsim/_core/dispatch.py
@@ -39,7 +39,7 @@ from smartsim.types import LaunchedJobID
 
 if t.TYPE_CHECKING:
     from smartsim._core.arguments.shell import ShellLaunchArguments
-    from smartsim._core.utils.launcher import ExecutableProtocol, LauncherProtocol
+    from smartsim._core.utils.launcher import LauncherProtocol
     from smartsim.experiment import Experiment
     from smartsim.settings.arguments import LaunchArguments
 
@@ -66,7 +66,7 @@ a job
 FormatterType: TypeAlias = t.Callable[
     [
         _DispatchableT,
-        "ExecutableProtocol",
+        t.Union[str, t.Sequence[str]],
         WorkingDirectory,
         EnvironMappingType,
         pathlib.Path,
@@ -78,7 +78,7 @@ FormatterType: TypeAlias = t.Callable[
 capable of being launched by a launcher.
 """
 _LaunchConfigType: TypeAlias = """_LauncherAdapter[
-        ExecutableProtocol,
+        t.Union[str, t.Sequence[str]],
         WorkingDirectory,
         EnvironMappingType,
         pathlib.Path,
@@ -271,7 +271,7 @@ class _DispatchRegistration(t.Generic[_DispatchableT, _LaunchableT]):
             )
 
         def format_(
-            exe: ExecutableProtocol,
+            exe: t.Union[str, t.Sequence[str]],
             path: pathlib.Path,
             env: EnvironMappingType,
             out: pathlib.Path,
