+++ tests/temp_tests/test_launchable.py
@@ -100,7 +100,7 @@ def test_job_init_deepcopy():
 
 
 def test_add_mpmd_pair():
-    entity = SmartSimEntity("test_name")
+    entity = SmartSimEntity("test_name", "echo", None, None)
 
     mpmd_job = MPMDJob()
     mpmd_job.add_mpmd_pair(entity, LaunchSettings("slurm"))
@@ -171,10 +171,10 @@ def test_add_mpmd_pair_check_launcher_error():
     """Test that an error is raised when a pairs is added to an mpmd
     job using add_mpmd_pair that does not have the same launcher type"""
     mpmd_pairs = []
-    entity1 = SmartSimEntity("entity1")
+    entity1 = SmartSimEntity("entity1", "echo", None, None)
     launch_settings1 = LaunchSettings("slurm")
 
-    entity2 = SmartSimEntity("entity2")
+    entity2 = SmartSimEntity("entity2", "echo", None, None)
     launch_settings2 = LaunchSettings("pals")
 
     pair1 = MPMDPair(entity1, launch_settings1)
