+++ tests/test_experiment.py
@@ -195,7 +195,7 @@ class EchoHelloWorldEntity(entity.SmartSimEntity):
     """A simple smartsim entity that meets the `ExecutableProtocol` protocol"""
 
     def __init__(self):
-        super().__init__("test-entity")
+        super().__init__("test-entity","echo", None, None)
 
     def __eq__(self, other):
         if type(self) is not type(other):
